refactor: use new message queue structure
